+++ eslint.config.js
@@ -22,16 +22,32 @@ export default tseslint.config(
     },
   },
   {
-    files: ['**/*.js', '**/*.mjs'],
+    files: ['**/*.js', '**/*.mjs', '**/*.cjs'],
     ...tseslint.configs.disableTypeChecked,
     languageOptions: {
       ...tseslint.configs.disableTypeChecked.languageOptions,
       globals: {
         Buffer: 'readonly',
         console: 'readonly',
+        module: 'readonly',
         process: 'readonly',
         URL: 'readonly',
       },
+    },
+  },
+  {
+    files: ['**/*.d.cts'],
+    ...tseslint.configs.disableTypeChecked,
+  },
+  {
+    files: ['**/*.cjs'],
+    languageOptions: {
+      globals: {
+        require: 'readonly',
+      },
+    },
+    rules: {
+      '@typescript-eslint/no-require-imports': 'off',
     },
   },
 );